refactor: dedupe error-listener boilerplate and repeater add-row tests
Per review-panel's simplification lens: the console/pageerror listener
setup was copy-pasted identically across 5 spec files, and the "Add New
Row" repeater test body was near-identically duplicated across 3.
Extracted captureConsoleErrors() and assertAddRowIncrementsRepeater()
into a shared error-capture.js. Also adds an explicit least-privilege
permissions block to the e2e-tests.yml workflow (security lens).
