+++ .github/workflows/e2e-tests.yml
@@ -6,6 +6,9 @@ on:
   pull_request:
     branches: [develop, main]
 
+permissions:
+  contents: read
+
 jobs:
   e2e:
     timeout-minutes: 60
