+++ Modules/Quotes/Tests/E2E/quotes.spec.js
@@ -1,6 +1,7 @@
 import { test, expect } from '../../../Core/Tests/E2E/test.js';
 import { tenantPath } from '../../../Core/Tests/E2E/tenant-path.js';
 import { assertRealListContent } from '../../../Core/Tests/E2E/list-assertions.js';
+import { assertAddRowIncrementsRepeater } from '../../../Core/Tests/E2E/error-capture.js';
 
 /**
  * Per the e2e-behavioral-testing skill: these tests prove the feature works,
@@ -73,24 +74,13 @@ test.describe('Quotes', () => {
   });
 
   test('"Add New Row" on the quote items repeater adds a real row, with no errors', async ({ page }) => {
-    /* Arrange */
-    const errors = [];
-    page.on('console', (msg) => { if (msg.type() === 'error') errors.push(msg.text()); });
-    page.on('pageerror', (err) => errors.push(err.message));
-
-    await page.goto(tenantPath('/quotes/create'));
+    /* Arrange, Act & Assert */
     // The "Quote items" section starts collapsed — the "Add New Row" button
     // doesn't exist in the DOM until it's expanded.
-    await page.getByRole('heading', { name: 'Quote items' }).click();
-    const addButton = page.getByRole('button', { name: 'Add New Row' });
-    await expect(addButton).toBeVisible();
-    const itemsBefore = await page.locator('.fi-fo-repeater-item').count();
-
-    /* Act */
-    await addButton.click();
-
-    /* Assert */
-    await expect(page.locator('.fi-fo-repeater-item')).toHaveCount(itemsBefore + 1);
-    expect(errors, `unexpected error(s) adding a quote item row:\n${errors.join('\n')}`).toHaveLength(0);
+    await assertAddRowIncrementsRepeater(page, {
+      createPath: '/quotes/create',
+      sectionHeading: 'Quote items',
+      itemLabel: 'quote item',
+    });
   });
 });
